Update send_transfer.py
Changed the file to match the java send_transaction test case a little. Using descriptors for what some of the parameter values are.
The original had two lines for the Address, I found that confusing and made it clear that it is the receiving address with validation and level 2 security level (I am not sure if using level 2 security is mandatory).